requirements.txt
numpy>=1.26
pandas>=2.0
scikit-learn>=1.2
matplotlib>=3.7
seaborn>=0.13
pillow>=10.0
streamlit>=1.24
